Split index.js into setup.js and destroy.js
